chore: pub auth_mysql
