+++ src/auth/src/lib.rs
@@ -21,7 +21,9 @@ mod user_provider;
 #[cfg(feature = "testing")]
 pub mod tests;
 
-pub use common::{user_provider_from_option, userinfo_by_name, HashedPassword, Identity, Password};
+pub use common::{
+    auth_mysql, user_provider_from_option, userinfo_by_name, HashedPassword, Identity, Password,
+};
 pub use error::{Error, Result};
 pub use permission::{PermissionChecker, PermissionReq, PermissionResp};
 pub use user_info::UserInfo;
